AI-teamHub-app/backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.29
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
pyjwt>=2.8
pwdlib[argon2]>=0.2
pydantic-settings>=2.0
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=0.23
anyio>=4.0
httpx>=0.27
ruff>=0.4
anthropic>=0.50
pypdf>=4.0
python-docx>=1.1
python-multipart>=0.0.9
